refactor(components): extract components from `cards` and create `contact` directory
