+++ src/components/welcome-to-my-page-card.tsx
@@ -1,8 +1,7 @@
 import { Card } from 'primereact/card'
 import ReactMarkdown from 'react-markdown'
 
-import { cardPassThrough } from '.'
-
+import { cardPassThrough } from '~/common/card-pass-through'
 import { useWelcomeToMyPage } from '~/hooks/api'
 
 export function WelcomeToMyPageCard() {
